Rename client module to api
This is because client and cli are very similar and is easier to get
confused.
