+++ src/greynoise/__init__.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from greynoise.client import GreyNoise  # noqa
+from greynoise.api import GreyNoise  # noqa
 
 __author__ = "GreyNoise Intelligence"
 __copyright__ = "Copyright, GreyNoise"
